requirements.txt
boto3==1.28.40
botocore==1.31.40
mysql_connector_repackaged==0.3.1
python_dateutil==2.8.2
Requests==2.31.0
schedule==1.2.0
